+++ src/lib.rs
@@ -47,7 +47,7 @@
 //!
 //! ```rust
 //! use m_captcha::{
-//!     cache::HashCache,
+//!     cache::{messages::VerifyCaptchaResult, HashCache},
 //!     master::{AddSiteBuilder, Master},
 //!     pow::{ConfigBuilder, Work},
 //!     system::SystemBuilder,
@@ -160,7 +160,26 @@
 //!     let res = system.verify_pow(payload.clone()).await;
 //!     assert!(res.is_ok());
 //!
-//!     Ok(())
+//!    // The client should submit the token to the mCaptcha protected service
+//!    // The service should validate the token received from the client
+//!    // with the mCaptcha server before processing client's
+//!    // request
+//!
+//!    // mcaptcha protected service sends the following paylaod to mCaptcha
+//!    // server:
+//!    let verify_msg = VerifyCaptchaResult {
+//!        token: res.unwrap(),
+//!        key: mcaptcha_name.into(),
+//!    };
+//!
+//!    // on mCaptcha server:
+//!    let res = system.validate_verification_tokens(verify_msg).await;
+//!    // mCaptcha will return true if token is valid and false if
+//!    // token is invalid
+//!    assert!(res.is_ok());
+//!    assert!(res.unwrap());
+//!
+//!    Ok(())
 //! }
 //! ```
 #![forbid(unsafe_code)]